requirements.txt
fastapi==0.115.0
geopy==2.4.1
googlemaps==4.10.0
numba==0.58.1
numpy
pandas==2.2.3
psycopg2_binary==2.9.9
pydantic==2.9.2
pyquadkey2==0.3.1
python-dotenv==1.0.1
scikit_learn==1.5.2
tqdm==4.66.1
gunicorn
uvicorn